+++ imod/mf6/model.py
@@ -19,10 +19,13 @@ from jinja2 import Template
 
 import imod
 from imod.logging import standard_log_decorator
+from imod.mf6.drn import Drainage
+from imod.mf6.ghb import GeneralHeadBoundary
 from imod.mf6.hfb import HorizontalFlowBarrierBase
 from imod.mf6.interfaces.imodel import IModel
 from imod.mf6.mf6_wel_adapter import Mf6Wel
 from imod.mf6.package import Package
+from imod.mf6.riv import River
 from imod.mf6.statusinfo import NestedStatusInfo, StatusInfo, StatusInfoBase
 from imod.mf6.utilities.mask import _mask_all_packages
 from imod.mf6.utilities.mf6hfb import merge_hfb_packages
@@ -36,6 +39,14 @@ from imod.typing import GridDataArray
 from imod.typing.grid import is_spatial_grid
 
 HFB_PKGNAME = "hfb_merged"
+SUGGESTION_TEXT = (
+    "-> You might fix this by calling the package's ``.cleanup()`` method."
+)
+PKGTYPES_WITH_CLEANUP = [River, Drainage, GeneralHeadBoundary, GridAgnosticWell]
+
+
+def pkg_has_cleanup(pkg: Package):
+    return any(isinstance(pkg, pkgtype) for pkgtype in PKGTYPES_WITH_CLEANUP)
 
 
 class Modflow6Model(collections.UserDict, IModel, abc.ABC):
@@ -237,7 +248,10 @@ class Modflow6Model(collections.UserDict, IModel, abc.ABC):
                 bottom=bottom,
             )
             if len(pkg_errors) > 0:
-                model_status_info.add(pkg_errors_to_status_info(pkg_name, pkg_errors))
+                footer = SUGGESTION_TEXT if pkg_has_cleanup(pkg) else None
+                model_status_info.add(
+                    pkg_errors_to_status_info(pkg_name, pkg_errors, footer)
+                )
 
         return model_status_info