+++ imod/tests/test_mf6/test_mf6_statusinfo.py
@@ -1,3 +1,5 @@
+from pytest import fixture
+
 from imod.mf6.statusinfo import NestedStatusInfo, StatusInfo
 
 
@@ -29,28 +31,67 @@ class TestStatusInfo:
         assert len(status_info.errors) == 3
 
 
+@fixture(scope="function")
+def nested_status_info():
+    root = NestedStatusInfo("root")
+    child1 = NestedStatusInfo("child1")
+    child2 = StatusInfo("child2")
+    grand_child1 = NestedStatusInfo("grandchild1")
+    grand_child2 = StatusInfo("grandchild2")
+
+    root.add(child1)
+    root.add(child2)
+
+    child1.add(grand_child1)
+    child1.add(grand_child2)
+
+    child2.add_error("test error1")
+    grand_child2.add_error("test error2")
+    grand_child2.add_error("test error3")
+    return root
+
+
+EXPECTED_TEXT_TEMPLATE = """\
+root:
+\t* child1:
+\t\t* grandchild1:
+\t\t* grandchild2:
+\t\t\t* test error2
+\t\t\t* test error3
+\t* child2:
+\t\t* test error1"""
+
+
 class TestNestedStatusInfo:
-    def test_errors_multiple_children_and_grandchildren_flatten_nested_errors(self):
-        # Arrange.
-        root = NestedStatusInfo()
-        child1 = NestedStatusInfo()
-        child2 = StatusInfo()
-        grand_child1 = NestedStatusInfo()
-        grand_child2 = StatusInfo()
-
-        root.add(child1)
-        root.add(child2)
-
-        child1.add(grand_child1)
-        child1.add(grand_child2)
-
-        child2.add_error("test error1")
-        grand_child2.add_error("test error2")
-
+    def test_flatten_nested_errors(self, nested_status_info):
         # Act.
-        has_errors = root.has_errors()
-        errors = root.errors
-
+        has_errors = nested_status_info.has_errors()
+        errors = nested_status_info.errors
         # Assert.
         assert has_errors
-        assert len(errors) == 2
+        assert len(errors) == 3
+
+    def test_set_footer(self, nested_status_info: NestedStatusInfo):
+        # Act
+        nested_status_info.set_footer_text("footer")
+
+        # Assert
+        nested_status_info._NestedStatusInfo__footer_text == "footer"
+
+    def test_to_string(self, nested_status_info):
+        # Arrange
+        expected_text = EXPECTED_TEXT_TEMPLATE.format(footer="")
+        # Act
+        actual = nested_status_info.to_string()
+        # Assert
+        assert expected_text == actual
+
+    def test_to_string_with_footer(self, nested_status_info):
+        # Arrange
+        footer_text = "footer"
+        expected_text = EXPECTED_TEXT_TEMPLATE + "\n\t" + footer_text
+        # Act
+        nested_status_info.set_footer_text(footer_text)
+        actual = nested_status_info.to_string()
+        # Assert
+        assert expected_text == actual